requirement.txt
requests==2.28.2
python-jose==3.3.0
fastapi==0.89.1
uvicorn==0.20.0
